requirements.txt
azure-functions
beautifulsoup4
atlassian-python-api
azure-storage-blob